pyproject.toml
[tool.poetry]
name = "workshop-02-aovivo"
version = "0.1.0"
description = "Qualidade em projeto de dados"
authors = ["A-Quaglia <augustoquaglia@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mkdocs = "^1.5.3"
mkdocs-mermaid2-plugin = "^1.1.1"
mkdocs-material = "^9.5.15"
mkdocstrings = {extras = ["python"], version = "^0.24.1"}
taskipy = "^1.12.2"
isort = "^5.13.2"
black = "^24.3.0"
pytest = "^8.1.1"
pandera = {extras = ["io"], version = "^0.18.3"}
sqlalchemy = "^2.0.29"
schema = "^0.7.5"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
duckdb = "^0.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
format = """
isort .
black .
"""
kill = "kill -9 $(lsof -t -i :8000)"
test = "pytest -v"
run = """
python3 app/main.py
"""
doc = "mkdocs serve"